Changed to frontend admin.

+++ tachikoma-grpc/src/main/kotlin/com/sourceforgery/tachikoma/incomingemailaddress/IncomingEmailAddressService.kt
@@ -43,7 +43,7 @@ private constructor(
     }
 
     fun deleteIncomingEmailAddress(request: IncomingEmailAddress) {
-        authentication.requireFrontend()
+        authentication.requireFrontendAdmin()
         val authenticationDBO = authenticationDAO.getActiveById(authentication.authenticationId)!!
 
         incomingEmailAddressDAO.delete(
